# Properly specialize nullif for scalar (4x faster)