--- a/datafusion/functions/Cargo.toml
+++ b/datafusion/functions/Cargo.toml
@@ -114,6 +114,10 @@ required-features = ["datetime_expressions"]
 
 [[bench]]
 harness = false
+name = "nullif"
+
+[[bench]]
+harness = false
 name = "date_bin"
 required-features = ["datetime_expressions"]