--- a/datafusion/functions/Cargo.toml
+++ b/datafusion/functions/Cargo.toml
@@ -115,6 +115,7 @@ required-features = ["datetime_expressions"]
 [[bench]]
 harness = false
 name = "nullif"
+required-features = ["core_expressions"]
 
 [[bench]]
 harness = false