@simonvandel 2024-07-29
In the first commmits of this PR, I saw a 4x perf increase, but CI pointed out that a test failed. It turned out the benchmark returned Error, which obviously was faster than doing any computation.

The error was `arrorw-select::nullif` not being able to handle scalar as first argument. After fixing the error by passing in a array of same size as rhs, the performance increase is much smaller (4%).

So in the end this PR might not give not benefit - feel free to close it if you want.