@simonvandel 2024-08-04
> Any chance you have a moment to file a ticket upstream in arrow-rs? If not, I will do so

I created https://github.com/apache/arrow-rs/issues/6193